requirements.txt
pandas==0.23.4
numpy==1.15.4
retry==0.9.2
requests==2.20.0
redis==3.2.0
python_dateutil==2.8.0
mysql_connector_repackaged==0.3.1
